requirements.txt
Django==4.1.7
pytz
djangorestframework
django-rest-passwordreset
psycopg2-binary
pyyaml
ujson
requests
celery
redis
gjango-filter
pytest
pytest-django
pytest-cov